naumer_ICML_2022

Overview

The repository naumer_ICML_2022 contains code related to my anticipated ICML submission in 2022.

The code is split into two main sections.

The folder src represents a library which represents the main building blocks of the work and could be applied to other problems.

The folder PlotGeneration contains the scripts to reproduce the plots in the manuscript.

This library is not registered through the Julia package registry, but the package manager can still install it through:

] add https://github.com/helmuthn/naumer_Dimensionality_2022.jl

Random Sampling

In our approximation of the value function, we need to generate random positive definite matrices. In this section, we include helper functions to generate samples from the appropriate distributions.

naumer_Dimensionality_2022.randomPSD — Function

randomPSD([rng = GLOBAL_RNG,] n, λ)

Generate a random n×n positive definite matrix with i.i.d. exponentially distributed eigenvalues.

Arguments

n: Dimensionality of the matrix

λ: Exponential distribution parameter

Returns

A random n×n matrix

naumer_Dimensionality_2022.samplePSD — Function

samplePSD([rng = GLOBAL_RNG,] K, n, λ)

Generate k random n×n positive definite matrices with i.i.d. exponentially distributed eigenvalues.

Arguments

K - Number of samples

n - Dimensionality of the matrix

λ - Exponential distribution parameter

Returns

An n×n×K array representing K random matrices

naumer_Dimensionality_2022.sampleStateSpace — Function

sampleStateSpace(system::AbstractSystem, trajectorySampleCount, timestepCount, burnin, τ)

Generates stateSampleCount random samples of the statespace. For now, initializes according to a high-variance Gaussian distribution, then steps the system forward in time to shape the density according to the system.

A small i.i.d. Gaussian random vector is added after each step to avoid degeneracy in value iteration.

Arguments

system - Chosen dynamical system

trajectorySampleCount - Number of trajectory samples

timestepCount - Number of timesteps per trajectory

'burnin` - Initial time advancement

τ - timestep size

Returns

A 2D array representing trajectorySampleCount * timestepCount samples, where columns represent the state.

Interpolation

This section includes functions involved in our local average interpolation.

naumer_Dimensionality_2022.min_dist — Function

min_dist(v, D)

Finds the minimum distance vector in D from v. Allows multidimensional array representation, assumes last axis of D indexes the vectors.

Arguments

v - Vector being compared

D - Set of vectors

Returns

(out, out_ind)

Where out is the closest vector, and out_ind is the position.

naumer_Dimensionality_2022.nearestNeighbor — Function

nearestNeighbor(target, samples, values)

Approximates the value at target with the value at the closest known sample.

Arguments

target - Point being evaluated, domain of function

samples - Known sample locations, domain of function

values - Values at known locations, range of function

Returns

Approximate value at target

naumer_Dimensionality_2022.buildNearestNeighbor — Function

buildNearestNeighbor(f, samples)

Evaluates a function f at a set of chosen locations to build the nearest neighbor approximation. Glorified broadcast function.

The last axis of samples is assumed to index the sample locations.

Arguments

f - The function to be approximated

samples- Set of representation points

Returns

An array of values at the given sample locations.

naumer_Dimensionality_2022.localAverage — Function

localAverage(targetPSD, targetState, psdSamples, stateSamples, values, d_max)

Computes the local average interpolation, defaulting to a nearest neighbor interpolation if there are no points within d_max of the desired target location.

The last axis of samples is assumed to index the sample locations.

Arguments

targetPSD - PSD point being evaluated, domain of function

targetState - State point being evaluated, domain of function

psdSamples - Known PSD sample locations, domain of function

stateSamples - Known state sample locations, domain of function

values - Values at known locations, range of function

d_max - Maximum distance to average

Returns

A local average approximation at a given location

Notes

values is indexed as ...

naumer_Dimensionality_2022.localAverageWeights — Function

localAverageWeights(targetPSD, targetState, psdSamples, stateSamples, values, d_max)

Computes the local average interpolation, defaulting to a nearest neighbor interpolation if there are no points within d_max of the desired target location.

The last axis of samples is assumed to index the sample locations.

Arguments

targetPSD - PSD point being evaluated, domain of function

targetState - State point being evaluated, domain of function

psdSamples - Known PSD sample locations, domain of function

stateSamples - Known state sample locations, domain of function

d_max - Maximum distance to average

Returns

(indices, weights)

Notes

values is indexed as ...

POMDP Tools

naumer_Dimensionality_2022.updateCRLB — Function

updateCRLB(   crlb, 
              action, 
              jacobian,
              σ²)

Updates the Fisher information based on the Jacobian of the flow and the current measurement under Gaussian noise.

Arguments

crlb - Previous Fisher information matrix

action - Measurement vector

jacobian - Jacobian of flow for current timestep

σ² - Measurement variance

Returns

Updated CRLB

naumer_Dimensionality_2022.optimalAction_NearestNeighbor — Function

optimalAction_NearestNeighbor(  crlb, actionSpace, 
                                samples, values, 
                                jacobian, σ²)

Computes the optimal action for a given state based on the current nearest neighbor approximation of the value function. Assumes linear measurement under additive Gaussian noise.

Arguments

crlb - Current Cramér-Rao bound

actionSpace- Action Space (finite)

samples - Sample locations in value function

values - Sample evaluations of value function

jacobian - Jacobian of flow for current timestep

σ² - Measurement variance

Returns

The action that minimizes the Cramér-Rao bound

naumer_Dimensionality_2022.valueUpdate_NearestNeighbor — Function

valueUpdate_NearestNeighbor(   crlb, 
                               γ, 
                               jacobian, 
                               actionSpace, 
                               samples, values)

Iterates the discounted Bellman equation for minimizing the CRLB at a given current Fisher information under the nearest neighbor value approximation for a given point.

Assumes linear measurement under additive Gaussian noise.

Arguments

crlb - Current Cramér-Rao bound

γ - Discount factor

jacobian - Jacobian of flow for current timestep

actionSpace- Action Space (finite)

samples - Sample locations in value function

values - Sample evaluations of value function

σ² - Noise variance

Returns

The updated value function approximation evaluated at crlb

naumer_Dimensionality_2022.valueIterate_NearestNeighbor — Function

valueIterate_NearestNeighbor(γ, jacobian, actionSpace, samples, values)

Iterates the discounted Bellman equation for minimizing the CRLB at a given current Fisher information under the nearest neighbor value approximation.

Assumes linear measurement under additive Gaussian noise.

Arguments

γ - Discount factor

jacobian - Jacobian of flow for current timestep

actionSpace- Action Space (finite)

samples - Sample locations in value function

values - Sample evaluations of value function

σ² - Measurement variance

Returns

The updated value function approximation values for all points in samples.

Notes

This function is multithreaded, remember to give Julia multiple threads when launching with julia -t NTHREADS, where NTHREADS is the desired number of threads.

naumer_Dimensionality_2022.valueIterate_NearestNeighbor_precompute! — Function

valueIterate_NearestNeighbor_precompute!(   out, 
                                            values, 
                                            transitionMap::Array{Int,2}, 
                                            psdSamples, 
                                            γ, 
                                            stateSampleCount)

Completes one step of value iteration using precomputed transition maps.

Arguments

out - Output Array (new value function approximation)

values - Current value function approximation

transitionMap - Transitions of state/crlb based on the action

psdSamples - CRLB samples

γ - Discount Factor

stateSampleCount - Number of state-space samples (for indexing psdSamples)

Returns

Nothing

Notes

This function is multithreaded

naumer_Dimensionality_2022.optimalAction_LocalAverage — Function

optimalAction_LocalAverage(     crlb, newState, actionSpace, 
                                psdSamples, stateSamples, values, 
                                jacobian, σ², d_max)

Computes the optimal action for a given state based on the current nearest neighbor approximation of the value function. Assumes linear measurement under additive Gaussian noise.

Arguments

crlb - Current Cramér-Rao bound

newState - Updated system state

actionSpace - Action Space (finite)

psdSamples - Sample PSD locations in value function

stateSamples - Sample state locations in value function

values - Sample evaluations of value function

jacobian - Jacobian of flow for current timestep

σ² - Measurement variance

d_max - Maximum distance for local average

Returns

The action that minimizes the Cramér-Rao bound

naumer_Dimensionality_2022.valueUpdate_LocalAverage — Function

valueUpdate_LocalAverage(   crlb, 
                            newState,
                            γ, 
                            jacobian, 
                            actionSpace, 
                            psdSamples, stateSamples,
                            values,
                            σ²,
                            d_max)

Iterates the discounted Bellman equation for minimizing the CRLB at a given current Fisher information under the nearest neighbor value approximation for a given point.

Assumes linear measurement under additive Gaussian noise.

Arguments

crlb - Current Cramér-Rao bound

newState - Updated state value

γ - Discount factor

jacobian - Jacobian of flow for current timestep

actionSpace - Action Space (finite)

psdSamples - Sample PSD locations in value function

stateSamples - Sample state locations in value function

values - Sample evaluations of value function

σ² - Noise variance

d_max - Maximum distance for local average

Returns

The updated value function approximation evaluated at crlb

Missing docstring.

Missing docstring for valueIterate_LocalAverage. Check Documenter's build log for details.

Dynamic Programming Solvers

Finally, we have the full value iteration solvers for dynamic programming.

naumer_Dimensionality_2022.ValueFunctionApproximation_NearestNeighbor_precompute — Function

ValueFunctionApproximation_NearestNeighbor_precompute(  system::AbstractSystem, 
                                                        τ,
                                                        γ, 
                                                        actionSpace, 
                                                        λ = 1,
                                                        psdSampleCount = 1000,
                                                        trajectorySampleCount = 100
                                                        timestepSampleCount = 10,
                                                        σ² = 1,
                                                        max_iterations=50)

Approximates the value function in the discounted cost optimal sampling problem. Requires a model of the system dynamics, timestep, discount-factor, and a discrete action space.

Precomputes all state transitions for all actions. Uses significant memory in exchange for less expensive computation.

Computes max_iterations steps of value iteration.

Arguments

system - Representation of the dynamical system

τ - Timestep size

γ - Discount Factor

actionSpace - Vector of possible actions

λ - CRLB space sampling parameter

psdSampleCount - Number of covariance-matrix samples

trajectorySampleCount - Number of state-space trajectory samples

timestepSampleCount - Number of state-space timestep samples

σ² - Measurement variance

max_iterations - Maximum number of iterations

Returns

(values, psdSamples, stateSamples)

values - Output of value function at sample points

psdSamples - CRLB sample locations

stateSamples - State sample locations

Note

values is a 1D array indexed as crlbIndex + psdSampleCount * (stateIndex - 1)

naumer_Dimensionality_2022.NearestNeighbor_OptimalPolicy — Function

NearestNeighbor_OptimalPolicy(  state, 
                                crlb, 
                                system::AbstractSystem, 
                                σ²,
                                τ, 
                                values, 
                                psdSamples, 
                                stateSamples, 
                                actionSpace)

Given a nearest neighbor approximation method for the value function, as well as the current state and crlb, returns the optimal action.

Arguments

state - Current state of the system

crlb - Current CRLB

system - Represntation of the system

σ² - Measurement noise power

τ - Timestep size

values - Value function evaluated at samples

psdSamples - Positive Semidefinite matrix samples

`stateSamples - Samples in the state-space

actionSpace - Set of possible actions

Returns

(action, index, new_state, new_crlb)

action - The optimal action

index - The index of the action

naumer_Dimensionality_2022.ValueFunctionApproximation_LocalAverage — Function

ValueFunctionApproximation_LocalAverage(    system::AbstractSystem, 
                                            τ,
                                            γ, 
                                            actionSpace, 
                                            λ = 1,
                                            psdSampleCount = 1000,
                                            trajectorySampleCount = 100
                                            timestepSampleCount = 10,
                                            σ² = 1,
                                            max_iterations=50,
                                            d_max = 0.05)

Approximates the value function using a LocalAverage interpolation in the discounted cost optimal sampling problem. Requires a model of the system dynamics, timestep, discount-factor, and a discrete action space.

Computes max_iterations steps of value iteration.

Arguments

system - Representation of the dynamical system

τ - Timestep size

γ - Discount Factor

actionSpace - Vector of possible actions

λ - CRLB space sampling parameter

psdSampleCount - Number of covariance-matrix samples

trajectorySampleCount - Number of state-space trajectory samples

timestepSampleCount - Number of state-space timestep samples

σ² - Measurement variance

max_iterations - Maximum number of iterations

d_max -

Returns

(values, psdSamples, stateSamples)

values - Output of value function at sample points

psdSamples - CRLB sample locations

stateSamples - State sample locations

Note

values is a 1D array indexed as crlbIndex + psdSampleCount * (stateIndex - 1)

naumer_Dimensionality_2022.LocalAverage_OptimalPolicy — Function

LocalAverage_OptimalPolicy( state, 
                            crlb, 
                            system::AbstractSystem, 
                            σ²,
                            τ, 
                            values, 
                            psdSamples, 
                            stateSamples, 
                            actionSpace,
                            d_max)

Given a nearest neighbor approximation method for the value function, as well as the current state and crlb, returns the optimal action.

Arguments

state - Current state of the system

crlb - Current CRLB

system - Represntation of the system

σ² - Measurement noise power

τ - Timestep size

values - Value function evaluated at samples

psdSamples - Positive Semidefinite matrix samples

`stateSamples - Samples in the state-space

actionSpace - Set of possible actions

d_max - Maximum distance for local average

Returns

(action, index, new_state, new_crlb)

action - The optimal action

index - The index of the action

naumer_Dimensionality_2022.ValueFunctionApproximation_LocalAverage_precompute — Function

ValueFunctionApproximation_LocalAverage_precompute(    system::AbstractSystem, 
                                                       τ,
                                                       γ, 
                                                       actionSpace, 
                                                       λ = 1,
                                                       psdSampleCount = 1000,
                                                       trajectorySampleCount = 100
                                                       timestepSampleCount = 10,
                                                       σ² = 1,
                                                       max_iterations=50,
                                                       d_max = 0.05)

Approximates the value function using a LocalAverage interpolation in the discounted cost optimal sampling problem. Requires a model of the system dynamics, timestep, discount-factor, and a discrete action space.

Computes max_iterations steps of value iteration.

Precomputes the local averaging weights to accelerate the value iteration steps.

Arguments

system - Representation of the dynamical system

τ - Timestep size

γ - Discount Factor

actionSpace - Vector of possible actions

λ - CRLB space sampling parameter

psdSampleCount - Number of covariance-matrix samples

trajectorySampleCount - Number of state-space trajectory samples

timestepSampleCount - Number of state-space timestep samples

σ² - Measurement variance

max_iterations - Maximum number of iterations

d_max -

Returns

(values, psdSamples, stateSamples)

values - Output of value function at sample points

psdSamples - CRLB sample locations

stateSamples - State sample locations

Note

values is a 1D array indexed as crlbIndex + psdSampleCount * (stateIndex - 1)

Extended Kalman Filter

We include a function for the extended Kalman filter under linear functional measurements.

naumer_Dimensionality_2022.stateupdate_EKF — Function

stateupdate_EKF(prediction, covariance, observation, action, σ²)

Updates the state based on the measurement according to the Extended Kalman Filter (EKF).

Arguments

prediction - Predicted State

covariance - Covariance approximation

observation - Observed value

action - Measurement functional

σ² - Noise variance

Returns

The updated state vector

1D Approximation

naumer_Dimensionality_2022.optimalaction_1DApprox — Function

optimalaction_1DApprox(actionspace, limitvector::Vector, crlb, σ²)

Returns the optimal action out of a finite actionspace.

Arguments

actionspace - List of measurement functionals

limitvector - Non-zero right singular vector of limiting flow Jacobian

crlb - Current CRLB for state estimation

σ² - Measurement noise variance

Returns

`index, action`

index - The index of the optimal action

action - The optimal action

optimalaction_1DApprox(actionspace, system::AbstractSystem, state, crlb, σ², T)

Returns the optimal action out of a finite actionspace.

Arguments

actionspace - List of measurement functionals

system - Dynamical System

state - Current system state

crlb - Current CRLB for state estimation

σ² - Measurement noise variance

T - Time horizon to approximate limit

Returns

`index, action`

index - The index of the optimal action

action - The optimal action

naumer_Dimensionality_2022.measurementvalue_1DApprox — Function

measurementvalue_1DApprox(action, limitvector::Vector, crlb, σ²)

Gives a value for the measurement action based on the 1D approximation of the limiting behavior.

Arguments

action - Measurement functional

limitvector - Non-zero right singular vector of limiting flow Jacobian

crlb - Current CRLB for state estimation

σ² - Measurement noise variance

Returns

A value of the action which should be maximized to minimize the CRLB for prediction.

naumer_Dimensionality_2022.actiongradientstep_1DApprox — Function

actiongradientstep_1DApprox(action, limitvector, crlb, σ², τ)

Gradient step along action space.

System API

To add the new abstract system, create a new subtype of AbstractSystem.

naumer_Dimensionality_2022.differential — Function

differential(system::AbstractSystem, x::Vector)

Computes the time derivative of a given continuous-time dynamical system evaluated at a given point.

Arguments

system - Continuous-Time Dynamical System

x - Point to Evaluate differential

Returns

The derivative of the trajectory of the system with respect to time evaluated at a fixed point x.

naumer_Dimensionality_2022.flow — Function

flow(x::Vector, τ, system::AbstractSystem)

Advances the state of a dynamical system from state x by time τ.

Arguments

system - Continuous-Time Dynamical System

τ - Advancement time

x - Initial state

Returns

The state advanced by τ units of time

naumer_Dimensionality_2022.flowJacobian — Function

flowJacobian(x::Vector, τ, system::AbstractSystem)

Computes the Jacobian of the flow of a system for τ units of time around an initial condition x.

Arguments

system - Continuous-Time Dynamical System

τ - Advancement time

x - Initial State

Returns

Jacobian matrix of the flow representing the derivative with respect to the initial state x.

naumer_Dimensionality_2022.dimension — Function

dimension(system::AbstractSystem)

Returns the dimensionality of the system.

Example Systems

We include a number of example systems in this work.

naumer_Dimensionality_2022.StaticSystem — Type

StaticSystem <: AbstractSystem

Defines a system dx/dt = 0

naumer_Dimensionality_2022.LinearSystem — Type

LinearSystem{T} <: AbstractSystem{T}

Defines a linear system dx/dt = Ax

Fields

dynamics::Matrix{T} - Matrix defining dynamics

naumer_Dimensionality_2022.VanDerPolSystem — Type

VanDerPolSystem{T} <: AbstractSystem{T}

naumer_Dimensionality_2022.VanDerPolSystem_expanded — Type

VanDerPolSystem_expanded

Represents dynamics in a product space where two axis represent a Van Der Pol System, while the rest represent stable linear systems.

Used to demonstrate scaling of dimensionality collapse benefit by allowing the introduction of arbitrary numbers of collapsing dimensions.

Fields

μ - Van Der Pol Parameter

linearDimension - Extra Collapsing Dimension Count

naumer_Dimensionality_2022.HopfSystem — Type

HopfSystem{T} <: AbstractSystem{T}

naumer_Dimensionality_2022.LorenzSystem — Type

LorenzSystem{T} <: AbstractSystem{T}