naumer_ICML_2022

Overview

The repository naumer_ICML_2022 contains code related to my anticipated ICML submission in 2022.

The code is split into two main sections.

The folder src represents a library which represents the main building blocks of the work and could be applied to other problems.

The folder PlotGeneration contains the scripts to reproduce the plots in the manuscript.

This library is not registered through the Julia package registry, but the package manager can still install it through:

] add https://github.com/helmuthn/naumer_ICML_2022.jl

For now, below is a list of functions and docstrings. The documentation will be filled out more after the manuscript is complete.

naumer_ICML_2022.LinearSystem

naumer_ICML_2022.ValueFunctionApproximation_NearestNeighbor_precompute

naumer_ICML_2022.buildNearestNeighbor

naumer_ICML_2022.differential

naumer_ICML_2022.dimension

naumer_ICML_2022.flow

naumer_ICML_2022.flowJacobian

naumer_ICML_2022.min_dist

naumer_ICML_2022.nearestNeighbor

naumer_ICML_2022.optimalAction_NearestNeighbor

naumer_ICML_2022.randomPSD

naumer_ICML_2022.samplePSD

naumer_ICML_2022.sampleStateSpace

naumer_ICML_2022.updateCRLB

naumer_ICML_2022.valueIterate_NearestNeighbor

naumer_ICML_2022.valueIterate_NearestNeighbor_precompute!

naumer_ICML_2022.valueUpdate_NearestNeighbor

naumer_ICML_2022.LinearSystem — Type

LinearSystem{T} <: AbstractSystem{T}

Defines a linear system dx/dt = Ax

Fields

dynamics::Matrix{T} - Matrix defining dynamics

naumer_ICML_2022.ValueFunctionApproximation_NearestNeighbor_precompute — Function

ValueFunctionApproximation_NearestNeighbor_precompute(  system::AbstractSystem, 
                                                        τ,
                                                        γ, 
                                                        actionSpace, 
                                                        λ = 1,
                                                        psdSampleCount = 1000,
                                                        trajectorySampleCount = 100
                                                        timestepSampleCount = 10,
                                                        σ² = 1,
                                                        max_iterations=50)

Approximates the value function in the discounted cost optimal sampling problem. Requires a model of the system dynamics, timestep, discount-factor, and a discrete action space.

Precomputes all state transitions for all actions. Uses significant memory in exchange for less expensive computation.

Computes max_iterations steps of value iteration.

Arguments

system - Representation of the dynamical system

τ - Timestep size

γ - Discount Factor

actionSpace - Vector of possible actions

λ - CRLB space sampling parameter

psdSampleCount - Number of covariance-matrix samples

trajectorySampleCount - Number of state-space trajectory samples

timestepSampleCount - Number of state-space timestep samples

σ² - Measurement variance

max_iterations - Maximum number of iterations

Returns

(values, psdSamples, stateSamples)

values - Output of value function at sample points

psdSamples - CRLB sample locations

stateSamples - State sample locations

Note

values is a 1D array indexed as crlbIndex + psdSampleCount * (stateIndex - 1)

naumer_ICML_2022.buildNearestNeighbor — Method

buildNearestNeighbor(f, samples)

Evaluates a function f at a set of chosen locations to build the nearest neighbor approximation. Glorified broadcast function.

The last axis of samples is assumed to index the sample locations.

Arguments

f - The function to be approximated

samples- Set of representation points

Returns

An array of values at the given sample locations.

naumer_ICML_2022.differential — Function

differential(system::AbstractSystem, x::Vector)

Computes the time derivative of a given continuous-time dynamical system evaluated at a given point.

Arguments

system - Continuous-Time Dynamical System

x - Point to Evaluate differential

Returns

The derivative of the trajectory of the system with respect to time evaluated at a fixed point x.

naumer_ICML_2022.dimension — Function

dimension(system::AbstractSystem)

Returns the dimensionality of the system.

naumer_ICML_2022.flow — Function

flow(system::AbstractSystem, τ, x::Vector)

Advances the state of a dynamical system from state x by time τ.

Arguments

system - Continuous-Time Dynamical System

τ - Advancement time

x - Initial state

Returns

The state advanced by τ units of time

naumer_ICML_2022.flowJacobian — Function

flowJacobian(system::AbstractSystem, τ, x::Vector)

Computes the Jacobian of the flow of a system for τ units of time around an initial condition x.

Arguments

system - Continuous-Time Dynamical System

τ - Advancement time

x - Initial State

Returns

Jacobian matrix of the flow representing the derivative with respect to the initial state x.

naumer_ICML_2022.min_dist — Method

min_dist(v, D)

Finds the minimum distance vector in D from v. Allows multidimensional array representation, assumes last axis of D indexes the vectors.

Arguments

v - Vector being compared

D - Set of vectors

Returns

(out, out_ind)

Where out is the closest vector, and out_ind is the position.

Notes

This function is not exported

naumer_ICML_2022.nearestNeighbor — Method

nearestNeighbor(target, samples, values)

Approximates the value at target with the value at the closest known sample.

Arguments

target - Point being evaluated, domain of function

samples - Known sample locations, domain of function

values - Values at known locations, range of function

Returns

Approximate value at target

naumer_ICML_2022.optimalAction_NearestNeighbor — Method

optimalAction_NearestNeighbor(  crlb, actionSpace, 
                                samples, values, 
                                jacobian, σ²)

Computes the optimal action for a given state based on the current nearest neighbor approximation of the value function. Assumes linear measurement under additive Gaussian noise.

Arguments

crlb - Current Cramér-Rao bound

actionSpace- Action Space (finite)

samples - Sample locations in value function

values - Sample evaluations of value function

jacobian - Jacobian of flow for current timestep

σ² - Measurement variance

Returns

The action that minimizes the Cramér-Rao bound

naumer_ICML_2022.randomPSD — Method

randomPSD([rng = GLOBAL_RNG,] n, λ)

Generate a random n×n positive definite matrix with i.i.d. exponentially distributed eigenvalues.

Arguments

n: Dimensionality of the matrix

λ: Exponential distribution parameter

Returns

A random n×n matrix

naumer_ICML_2022.samplePSD — Method

samplePSD([rng = GLOBAL_RNG,] K, n, λ)

Generate k random n×n positive definite matrices with i.i.d. exponentially distributed eigenvalues.

Arguments

K - Number of samples

n - Dimensionality of the matrix

λ - Exponential distribution parameter

Returns

An n×n×K array representing K random matrices

naumer_ICML_2022.sampleStateSpace — Method

sampleStateSpace(system::AbstractSystem, trajectorySampleCount, timestepCount, burnin, τ)

Generates stateSampleCount random samples of the statespace. For now, initializes according to a high-variance Gaussian distribution, then steps the system forward in time to shape the density according to the system.

A small i.i.d. Gaussian random vector is added after each step to avoid degeneracy in value iteration.

Arguments

system - Chosen dynamical system

trajectorySampleCount - Number of trajectory samples

timestepCount - Number of timesteps per trajectory

'burnin` - Initial time advancement

τ - timestep size

Returns

A 2D array representing trajectorySampleCount * timestepCount samples, where columns represent the state.

naumer_ICML_2022.updateCRLB — Method

updateFisherInformation(    crlb, 
                            action, 
                            jacobian,
                            σ²)

Updates the Fisher information based on the Jacobian of the flow and the current measurement under Gaussian noise.

Arguments

crlb- Previous Fisher information matrix

action - Measurement vector

jacobian - Jacobian of flow for current timestep

σ² - Measurement variance

Returns

Updated Fisher Information

naumer_ICML_2022.valueIterate_NearestNeighbor — Method

valueIterate_NearestNeighbor(γ, jacobian, actionSpace, samples, values)

Iterates the discounted Bellman equation for minimizing the CRLB at a given current Fisher information under the nearest neighbor value approximation.

Assumes linear measurement under additive Gaussian noise.

Arguments

γ - Discount factor

jacobian - Jacobian of flow for current timestep

actionSpace- Action Space (finite)

samples - Sample locations in value function

values - Sample evaluations of value function

σ² - Measurement variance

Returns

The updated value function approximation values for all points in samples.

Notes

This function is multithreaded, remember to give Julia multiple threads when launching with julia -t NTHREADS, where NTHREADS is the desired number of threads.

naumer_ICML_2022.valueIterate_NearestNeighbor_precompute! — Method

valueIterate_NearestNeighbor_precompute!(out, values, transitionMap, psdSamples, γ, stateSampleCount)

Completes one step of value iteration using precomputed transition maps.

Arguments

out - Output Array (new value function approximation)

values - Current value function approximation

transitionMap - Transitions of state/crlb based on the action

psdSamples - CRLB samples

γ - Discount Factor

stateSampleCount - Number of state-space samples (for indexing psdSamples)

Returns

Nothing

Notes

This function is multithreaded

naumer_ICML_2022.valueUpdate_NearestNeighbor — Method

valueUpdate_NearestNeighbor(   crlb, 
                               γ, 
                               jacobian, 
                               actionSpace, 
                               samples, values)

Iterates the discounted Bellman equation for minimizing the CRLB at a given current Fisher information under the nearest neighbor value approximation for a given point.

Assumes linear measurement under additive Gaussian noise.

Arguments

crlb - Current Cramér-Rao bound

γ - Discount factor

jacobian - Jacobian of flow for current timestep

actionSpace- Action Space (finite)

samples - Sample locations in value function

values - Sample evaluations of value function

σ² - Noise variance

Returns

The updated value function approximation evaluated at crlb