naumer_ICML_2022

Documentation for naumer_ICML_2022.

naumer_ICML_2022.nearestNeighbor

naumer_ICML_2022.optimalAction_NearestNeighbor

naumer_ICML_2022.randomPSD

naumer_ICML_2022.updateFisherInformation

naumer_ICML_2022.valueIterate

naumer_ICML_2022.valueUpdate_NearestNeighbor

naumer_ICML_2022.nearestNeighbor — Method

nearestNeighbor(target, samples, values)

Approximates the value at target with the value at the closest known sample.

Arguments

target - Point being evaluated, domain of function

samples - Known sample locations, domain of function

values - Values at known locations, range of function

Returns

Approximate value at target

naumer_ICML_2022.optimalAction_NearestNeighbor — Method

optimalAction_NearestNeighbor(  information, actionSpace, 
                                samples, values, 
                                jacobian)

Computes the optimal action for a given state based on the current nearest neighbor approximation of the value function.

Arguments

information- Current Fisher information

actionSpace- Action Space (finite)

samples - Sample locations in value function

values - Sample evaluations of value function

jacobian - Jacobian of flow for current timestep

Returns

The action that minimizes the Cramér-Rao bound

naumer_ICML_2022.randomPSD — Function

randomPSD([rng,] n, λ=1)

Generate a random n×n positive definite matrix with i.i.d. exponentially distributed eigenvalues.

Arguments

n: Dimensionality of the matrix

λ: Exponential distribution parameter

Returns

A random n×n matrix

naumer_ICML_2022.updateFisherInformation — Method

updateFisherInformation(    information, 
                            action, 
                            jacobian,
                            σ²)

Updates the Fisher information based on the Jacobian of the flow and the current measurement under Gaussian noise.

Arguments

information- Previous Fisher information matrix

action - Measurement vector

jacobian - Jacobian of flow for current timestep

σ² - Measurement variance

Returns

Updated Fisher Information

naumer_ICML_2022.valueIterate — Method

ValueIterate_NearestNeighbor(γ, jacobian, actionSpace, samples, values)

Iterates the discounted Bellman equation for minimizing the CRLB at a given current Fisher information under the nearest neighbor value approximation.

Arguments

γ - Discount factor

jacobian - Jacobian of flow for current timestep

actionSpace- Action Space (finite)

samples - Sample locations in value function

values - Sample evaluations of value function

Returns

The updated value function approximation values for all points in samples.

Notes

This function is multithreaded, remember to give Julia multiple threads when launching with julia -t NTHREADS, where NTHREADS is the desired number of threads.

naumer_ICML_2022.valueUpdate_NearestNeighbor — Method

valueUpdate_NearestNeighbor(   information, 
                               γ, 
                               jacobian, 
                               actionSpace, 
                               samples, values)

Iterates the discounted Bellman equation for minimizing the CRLB at a given current Fisher information under the nearest neighbor value approximation for a given point.

Arguments

information- Current Fisher information

γ - Discount factor

jacobian - Jacobian of flow for current timestep

actionSpace- Action Space (finite)

samples - Sample locations in value function

values - Sample evaluations of value function

Returns

The updated value function approximation evaluated at information